Move BeansXmlMerging tests into environments/se/test folder to prevent classpath pollution
